RE-461 Removes Kafka client library from Dockerfile
Have followed up with @chriswhunter89 but this doesn't seem to need to be in here, at least not on a permanent basis. From what I can see it maybe it time was used to write something out to the mounted PVC.

If I've got this correct and the above is right then we should be getting what is in the PVC and saving it as a Kubernetes Secret for this to mount.